Do not set a timer to ping currently unavailable devices
The documentation of availability_timeout says: "When enabled, devices
will be checked if they are still online.". So this actually makes it
work as the documentation says.

Also, unavailable devices will anyway announce themselves when they are
available again (at least my hue lights and innr plug do that),
so it's only neccesary to ping available devices periodically to
know when they become unavailable.

In an installation with many offline devices, this can reduce the
zigbee traffic.