Add randomize_availability_timeout setting
Add a new randomize_availability_timeout setting that allows to
randomize the timeout of availability checks so all checks are not sent
at once.

Instead of using a fixed timeout, a random number of seconds between 0
and randomize_availability_timeout is added to the usual timeout each
time the timer is set, so we separate the pings a random time, so we
don't saturate the zigbee network.