feat: Implement systemd-notify directly
Dependency on sd-notify removed.
sd-notify is wrapping libsystemd, which requires installing
libsystemd-dev for building, and loads libsystemd.so at runtime.
Since we only need the sd_notify part of the lib, and the protocol
is trivial to implement, do so.

We also now bail out early in case a watchdog timeout is set and
the notification module could not be loaded, as in this case
systemd would constantly kill and restart zigbee2mqtt, resulting
in unstable operation that might not be obvious at first sight.